Extracts RGB out of Bubble
