+++ Borkle/Borkle/Bubble.swift
@@ -21,40 +21,6 @@ class Bubble: Codable {
         }
     }
 
-    struct RGB: Codable {
-        let red: CGFloat
-        let green: CGFloat
-        let blue: CGFloat
-
-        init(red: CGFloat, green: CGFloat, blue: CGFloat) {
-            self.red = red
-            self.green = green
-            self.blue = blue
-        }
-
-        init(string: String?) {
-            guard let string = string else {
-                // no string, be obnoxious green
-                red = 0.0; green = 1.0; blue = 0.0
-                return
-            }
-
-            let chunks = string
-              .split(separator: " ")
-              .compactMap { String($0) }
-              .compactMap { CGFloat($0) }
-
-            guard chunks.count >= 3 else {
-                // not enough chunkage, be obnoxious green
-                red = 0.0; green = 1.0; blue = 0.0
-                return
-            }
-            red = chunks[0]
-            green = chunks[1]
-            blue = chunks[2]
-        }
-    }
-
     var fillColorRGB: RGB?
     var fillColor: NSColor? {
         get {
@@ -265,17 +231,17 @@ extension Bubble {
 }
 
 extension NSColor {
-    static func colorFromRGB(_ rgb: Bubble.RGB) -> NSColor {
+    static func colorFromRGB(_ rgb: RGB) -> NSColor {
         NSColor.init(deviceRed: rgb.red, green: rgb.green, blue: rgb.blue, alpha: 1.0)
     }
 
-    func rgbColor() -> Bubble.RGB {
+    func rgbColor() -> RGB {
         var red: CGFloat = 0.0
         var green: CGFloat = 0.0
         var blue: CGFloat = 0.0
 
         getRed(&red, green: &green, blue: &blue, alpha: nil)
-        return Bubble.RGB(red: red, green: green, blue: blue)
+        return RGB(red: red, green: green, blue: blue)
     }
 
 }
