+++ Borkle/Borkle/ScappleImporter.swift
@@ -161,10 +161,10 @@ extension ScappleImporter: XMLParserDelegate {
             currentFormattingOptionsAttributes = nil
             currentFormattingOptionsString = nil
         case "Border":
-            currentBubble.borderColorRGB = Bubble.RGB(string: currentBorderColorString)
+            currentBubble.borderColorRGB = RGB(string: currentBorderColorString)
             currentBorderColorString = nil
         case "Fill":
-            currentBubble.fillColorRGB = Bubble.RGB(string: currentFillString)
+            currentBubble.fillColorRGB = RGB(string: currentFillString)
             currentFillString = nil
         default:
             break
